requirements.txt
Django==4.0.2
requests==2.27.1
django-environ==0.8.1